+++ src/low-level/imap/mailimap_sender.c
@@ -1565,7 +1565,7 @@ mailimap_literal_data_send(mailstream * fd, const char * literal, uint32_t len,
 }
 
 int
-mailimap_literal_data_send_with_context(mailstream * fd, const char * literal, size_t len,
+mailimap_literal_data_send_with_context(mailstream * fd, const char * literal, uint32_t len,
                                         mailprogress_function * progr_fun,
                                         void * context)
 {
